Add utils file for sharing common code
