+++ constants.py
@@ -1,0 +1,1 @@
+RACKS = ["A", "B"]
